Move the component to the `notebook` directory
