+++ frontend/src/metabase/query_builder/components/notebook/NotebookNativePreview/utils.ts
